fix: remove old product components and remove all categoryreferences
