--- apps/sap-commerce-cloud/src/__mocks__/mockCategoryPreview.js
@@ -1,8 +1,0 @@
-export const mockCategoryPreview = {
-  sku: 'abc1234',
-  image: '',
-  id: '123',
-  name: 'Mock Product',
-  slug: '',
-  isMissing: false,
-};
