+++ apps/sap-commerce-cloud/src/__mocks__/mockProductPreview.ts
@@ -4,7 +4,7 @@ export const mockProductPreview: Product = {
   sku: 'abc1234',
   image: '',
   id: '123',
-  name: 'Mock Category',
+  name: 'Mock Product',
   isMissing: false,
   productUrl: '/product/abc1234',
 };
