--- apps/sap-commerce-cloud/src/__mocks__/react-sortable-hoc.js
@@ -1,3 +1,0 @@
-export const SortableContainer = (component) => component;
-export const SortableElement = (component) => component;
-export const SortableHandle = (component) => component;
