--- apps/sap-commerce-cloud/src/components/CategoryPreviews/SortableList.spec.tsx
@@ -1,22 +1,0 @@
-import { render, cleanup } from '@testing-library/react';
-import { Props, SortableList } from './SortableList';
-import { mockCategoryPreview } from '../../__mocks__/mockCategoryPreview';
-
-const defaultProps: Props = {
-  disabled: false,
-  categoryPreviews: [mockCategoryPreview],
-  deleteFn: jest.fn(),
-};
-
-const renderComponent = (props: Props) => {
-  return render(<SortableList {...props} />);
-};
-
-describe('SortableList', () => {
-  afterEach(cleanup);
-
-  it('should render successfully', async () => {
-    const component = renderComponent(defaultProps);
-    expect(component.container).toMatchSnapshot();
-  });
-});
