+++ apps/sap-commerce-cloud/src/components/Field.tsx
@@ -4,11 +4,8 @@ import { FieldAppSDK } from '@contentful/app-sdk';
 import { css } from 'emotion';
 import tokens from '@contentful/forma-36-tokens';
 import get from 'lodash/get';
-import { ProductPreviews } from './ProductPreviews/ProductPreviews';
-import { SortableComponent } from './ProductPreviewsV2/SortableComponent';
-import { CategoryPreviews } from './CategoryPreviews/CategoryPreviews';
+import { SortableComponent } from './ProductPreviews/SortableComponent';
 import { fetchProductPreviews } from '../api/fetchProductPreviews';
-import { fetchCategoryPreviews } from '../api/fetchCategoryPreviews';
 import logo from '../logo.png';
 import { AppParameters, SAPParameters } from '../interfaces';
 
@@ -43,16 +40,9 @@ function fieldValueToState(value?: string | string[]): string[] {
   return Array.isArray(value) ? value : [value];
 }
 
-function makeCTAText(fieldType: string, pickerMode: 'category' | 'product') {
+function makeCTAText(fieldType: string) {
   const isArray = fieldType === 'Array';
-  const beingSelected =
-    pickerMode === 'category'
-      ? isArray
-        ? 'categories'
-        : 'a category'
-      : isArray
-      ? 'products'
-      : 'a product';
+  const beingSelected = isArray ? 'products' : 'a product';
   return `Select ${beingSelected}`;
 }
 
@@ -76,18 +66,6 @@ export default class Field extends React.Component<Props, State> {
     });
   }
 
-  getPickerMode = () => {
-    const { sdk } = this.props;
-    const contentTypeId = sdk.contentType.sys.id;
-    const fieldId = sdk.field.id;
-
-    return get(
-      sdk,
-      ['parameters', 'installation', 'fieldsConfig', contentTypeId, fieldId],
-      'product'
-    );
-  };
-
   updateStateValue = (skus: string[]) => {
     this.setState({ value: skus });
 
@@ -105,7 +83,7 @@ export default class Field extends React.Component<Props, State> {
     const skus = await sdk.dialogs.openCurrentApp({
       allowHeightOverflow: true,
       position: 'center',
-      title: makeCTAText(sdk.field.type, this.getPickerMode()),
+      title: makeCTAText(sdk.field.type),
       shouldCloseOnOverlayClick: true,
       shouldCloseOnEscapePress: true,
       parameters: {
@@ -113,7 +91,6 @@ export default class Field extends React.Component<Props, State> {
         fieldValue: fieldValueToState(sdk.field.getValue()),
         fieldType: sdk.field.type,
         fieldId: sdk.field.id,
-        pickerMode: this.getPickerMode(),
       },
       width: 1400,
       minHeight: 600,
@@ -128,46 +105,22 @@ export default class Field extends React.Component<Props, State> {
   render = () => {
     const { value: data, editingDisabled } = this.state;
 
-    const isPickerTypeSetToCategory = this.getPickerMode() === 'category';
     const hasItems = data.length > 0;
-    const config = this.props.sdk.parameters;
     const fieldType = get(this.props, ['sdk', 'field', 'type'], '');
 
     return (
       <>
         {hasItems && (
           <div className={styles.sortable}>
-            {isPickerTypeSetToCategory ? (
-              <div></div>
-            ) : (
-              // <CategoryPreviews
-              //   sdk={this.props.sdk}
-              //   disabled={editingDisabled}
-              //   categories={data}
-              //   onChange={this.updateStateValue}
-              //   fetchCategoryPreviews={(categories) =>
-              //     fetchCategoryPreviews(categories, config.installation)
-              //   }
-              // />
-              // <ProductPreviews
-              //   sdk={this.props.sdk}
-              //   disabled={editingDisabled}
-              //   skus={data}
-              //   onChange={this.updateStateValue}
-              //   fetchProductPreviews={(skus) =>
-              //     fetchProductPreviews(skus, this.props.sdk.parameters as SAPParameters)
-              //   }
-              // />
-              <SortableComponent
-                sdk={this.props.sdk}
-                disabled={editingDisabled}
-                skus={data}
-                onChange={this.updateStateValue}
-                fetchProductPreviews={(skus) =>
-                  fetchProductPreviews(skus, this.props.sdk.parameters as SAPParameters)
-                }
-              />
-            )}
+            <SortableComponent
+              sdk={this.props.sdk}
+              disabled={editingDisabled}
+              skus={data}
+              onChange={this.updateStateValue}
+              fetchProductPreviews={(skus) =>
+                fetchProductPreviews(skus, this.props.sdk.parameters as SAPParameters)
+              }
+            />
           </div>
         )}
         <div className={styles.container}>
@@ -178,7 +131,7 @@ export default class Field extends React.Component<Props, State> {
             size="small"
             onClick={this.onDialogOpen}
             disabled={editingDisabled}>
-            {makeCTAText(fieldType, this.getPickerMode())}
+            {makeCTAText(fieldType)}
           </Button>
         </div>
       </>
