--- apps/sap-commerce-cloud/src/components/ProductPreviews/ProductPreviews.tsx
@@ -1,79 +1,0 @@
-import React from 'react';
-import arrayMove from 'array-move';
-import isEqual from 'lodash/isEqual';
-import difference from 'lodash/difference';
-import { FieldExtensionSDK } from '@contentful/app-sdk';
-import { mapSort } from '../../utils';
-import { SortableList } from './SortableList';
-import { PreviewsFn, Product } from '../../interfaces';
-
-interface Props {
-  sdk: FieldExtensionSDK;
-  disabled: boolean;
-  onChange: (skus: string[]) => void;
-  skus: string[];
-  fetchProductPreviews: PreviewsFn;
-}
-
-interface State {
-  productPreviews: Product[];
-}
-
-export class ProductPreviews extends React.Component<Props, State> {
-  state = {
-    productPreviews: [],
-  };
-
-  componentDidMount() {
-    this.updateProductPreviews();
-  }
-
-  componentDidUpdate({ skus: prevSKUs }: Props) {
-    if (!isEqual(this.props.skus, prevSKUs)) {
-      const lengthHasChanged = this.props.skus.length !== prevSKUs.length;
-      const skusDiffer = difference(this.props.skus, prevSKUs).length > 0;
-      const shouldRefetchProducts = lengthHasChanged || skusDiffer;
-      this.updateProductPreviews(shouldRefetchProducts);
-    }
-  }
-
-  updateProductPreviews = async (shouldRefetch: boolean = true) => {
-    try {
-      const { fetchProductPreviews, skus } = this.props;
-      const productPreviewsUnsorted = shouldRefetch
-        ? await fetchProductPreviews(skus)
-        : this.state.productPreviews;
-      const productPreviews = mapSort(productPreviewsUnsorted, skus, 'sku');
-      this.setState({ productPreviews });
-    } catch {
-      this.props.sdk.notifier.error(
-        'There was an error fetching the data for the selected products.'
-      );
-    }
-  };
-
-  onSortEnd = ({ oldIndex, newIndex }: { oldIndex: number; newIndex: number }) => {
-    const skus = arrayMove(this.props.skus, oldIndex, newIndex);
-    this.props.onChange(skus);
-  };
-
-  deleteItem = (index: number) => {
-    const skus = [...this.props.skus];
-    skus.splice(index, 1);
-    this.props.onChange(skus);
-  };
-
-  render() {
-    return (
-      <SortableList
-        disabled={this.props.disabled}
-        onSortStart={(_, e) => e.preventDefault()} // Fixes FF glitches.
-        onSortEnd={this.onSortEnd}
-        axis="xy"
-        productPreviews={this.state.productPreviews}
-        deleteFn={this.deleteItem}
-        useDragHandle
-      />
-    );
-  }
-}
