+++ apps/sap-commerce-cloud/src/components/ProductPreviews/SortableComponent.tsx
