+++ apps/sap-commerce-cloud/src/components/ProductPreviews/SortableList.tsx
@@ -1,5 +1,4 @@
-import * as React from 'react';
-import { SortableContainer } from 'react-sortable-hoc';
+import { FC } from 'react';
 import { DeleteFn, Product } from '../../interfaces';
 import { SortableListItem } from './SortableListItem';
 
@@ -9,24 +8,22 @@ export interface Props {
   deleteFn: DeleteFn;
 }
 
-export const SortableList = SortableContainer<Props>(
-  ({ disabled, deleteFn, productPreviews }: Props) => {
-    const itemsAreSortable = productPreviews.length > 1;
-    return (
-      <div>
-        {productPreviews.map((product, index) => {
-          return (
-            <SortableListItem
-              disabled={disabled}
-              key={`${product.image}-${product.sku}`}
-              product={product}
-              index={index}
-              onDelete={() => deleteFn(index)}
-              isSortable={itemsAreSortable}
-            />
-          );
-        })}
-      </div>
-    );
-  }
-);
+export const SortableList: FC<Props> = ({ disabled, deleteFn, productPreviews = [] }) => {
+  const itemsAreSortable = productPreviews.length > 1;
+
+  return (
+    <div>
+      {productPreviews.map((product, index) => {
+        return (
+          <SortableListItem
+            disabled={disabled}
+            key={`${product.image}-${product.sku}`}
+            product={product}
+            onDelete={() => deleteFn(index)}
+            isSortable={itemsAreSortable}
+          />
+        );
+      })}
+    </div>
+  );
+};
