--- apps/sap-commerce-cloud/src/components/ProductPreviewsV2/SortableList.spec.tsx
@@ -1,23 +1,0 @@
-import { render, cleanup, screen } from '@testing-library/react';
-import { Props, SortableList } from './SortableList';
-import { mockProductPreview } from '../../__mocks__';
-
-const defaultProps: Props = {
-  disabled: false,
-  productPreviews: [mockProductPreview],
-  deleteFn: jest.fn(),
-};
-
-const renderComponent = (props: Props) => {
-  return render(<SortableList {...props} />);
-};
-
-describe('SortableList', () => {
-  afterEach(cleanup);
-
-  it('should render successfully', async () => {
-    const { queryAllByTestId } = renderComponent(defaultProps);
-
-    expect(queryAllByTestId(`sortable-list-item`)).toHaveLength(1);
-  });
-});
